[#67] feat:type 공통 dto와 로그인 회원가입 dto 정의
